Changed component spacing and revised components' structure - Navigation, Footer: changed img to Image - Footer: added footer tag and added bg image, removed div with bg image - Home: added Bootstrap classes for spacing - Navigation: moved LoginModal to separate file, changed xs to md for Navbar, refactored code into variables - Overview: used p tags and changed xs to md for Col - Results: added Bootstrap class for spacing - Upload: added md={5} to Col - App: Moved Footer down so its Navbar doesn't get rendered next to the Navigation Navbar - index.html, index.tsx: used Bootstrap min CSS rather than link to external stylesheet
