+++ src/index.tsx
@@ -1,5 +1,6 @@
 import ReactDOM from "react-dom";
 import App from "./App";
 import "./css/index.css";
+import "bootstrap/dist/css/bootstrap.min.css";
 
 ReactDOM.render(<App />, document.getElementById("root"));
